refactor: Rename agent_settings.py to agent_preferences.py for consistency
Complete the preferences terminology alignment by renaming the LLM agent module from agent_settings.py to agent_preferences.py and updating all references:

- Rename src/areyouok_telegram/llms/agent_settings.py → agent_preferences.py
- Update imports in onboarding.py to use agent_preferences module
- Update mock patch decorators in tests to reference new module path
- Rename tests/test_llms/test_agent_settings.py → test_agent_preferences.py

This ensures complete consistency between user-facing commands (/preferences) and internal module naming.
