+++ tests/test_handlers/test_preferences.py
